+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+- Added `ProtocolVersion.V7` corresponding to Protocol version 7
+- Added `cooldowns` list to `AccountInfo`
+
 ## 7.2.0
 - Added `MessageSigningDigest` class to generate digests for message signing
 - Added support for company identity attributes